release: v0 0 2 herokuへのデプロイ準備

+++ app/Http/Controllers/RacketImageController.php
@@ -9,6 +9,7 @@ use App\Models\Maker;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 use Intervention\Image\Facades\Image;
+use Exception;
 
 class RacketImageController extends Controller
 {
@@ -46,46 +47,17 @@ class RacketImageController extends Controller
         $validated_request = $request->validated();
 
         try {
-            // 画像ファイルリサイジング
-            $file = Image::make($request->file('file'));
-            $file->orientate();
-            $file->resize(
-                480,
-                null,
-                function ($constraint) {
-                    // 縦横比を保持したままにする
-                    $constraint->aspectRatio();
-                    // 小さい画像は大きくしない
-                    $constraint->upsize();
-                }
-            );
-
-            $filename = now()->format('YmdHis') . $validated_request['title'] . "." . $request->file('file')->extension();
-
-            // storageに登録するためのpathを生成
-            $storagePath = storage_path('app/public/images/rackets');
-            $fileLocationFullPath = $storagePath . '/' . $filename;
-
-            if ($file->save($fileLocationFullPath)) {
-                // "/var/www/html/strii-backend/storage/app/public/images/rackets/20240105123954リサイズ確認５.jpg"
-                // intervension image導入前の登録の仕様に合わせるため
-                // 上記のようなfullPathをDBのfile_pathカラム用に整形
-                $trimedFilePath = strstr($fileLocationFullPath, 'images');
-
-                $racket_image = RacketImage::create([
-                    'file_path' => $trimedFilePath,
-                    'title' => $validated_request['title'],
-                    'maker_id' => $validated_request['maker_id']
-                ]);
-            }
+            $racket_image = RacketImage::registerRacketImage($validated_request);
 
             if (isset($racket_image)) {
                 $maker = Maker::find($racket_image->maker_id);
 
                 return response()->json([
-                    'file_path' => Storage::url($racket_image['file_path']),
+                    'id' => $racket_image['id'],
+                    'file_path' => $racket_image['file_path'],
                     'title' => $racket_image['title'],
-                    'maker' => $maker
+                    'maker' => $maker,
+                    'posting_user_id' => $racket_image['posting_user_id']
                 ], 200);
             }
         } catch (\Throwable $e) {
@@ -135,8 +107,8 @@ class RacketImageController extends Controller
         try {
             $image = RacketImage::with('maker')->findOrFail($id);
 
-            //新しいファイルがあれば新たにstorageに登録
-            if($request->file('file')) {
+            //新しいファイルがあれば新たにs3へアップロードする
+            if ($request->file('file')) {
                 // 画像ファイルリサイジング
                 $file = Image::make($request->file('file'));
                 $file->orientate();
@@ -152,39 +124,68 @@ class RacketImageController extends Controller
                 );
 
                 $filename = now()->format('YmdHis') . $validated_request['title'] . "." . $request->file('file')->extension();
-
-                // storageに登録するためのpathを生成
+                
+                // storageに一時的に登録するためのpathを生成
                 $storagePath = storage_path('app/public/images/rackets');
                 $fileLocationFullPath = $storagePath . '/' . $filename;
 
                 if ($file->save($fileLocationFullPath)) {
-                    // "/var/www/html/strii-backend/storage/app/public/images/rackets/20240105123954リサイズ確認５.jpg"
-                    // intervension image導入前の登録の仕様に合わせるため
-                    // 上記のようなfullPathをDBのfile_pathカラム用に整形
-                    $trimedFilePath = strstr($fileLocationFullPath, 'images');
+                    // 画像をAmazon S3にアップロード用urlに整形
+                    $filePath = 'images/rackets/' . $filename;
 
-                    //以前のイメージファイルをstorageフォルダから削除
-                    Storage::disk('public')->delete($image->file_path);
+                    // s3へアップロード
+                    if (Storage::disk('s3')->put($filePath, file_get_contents($fileLocationFullPath))) {
+                        // 成功時
+                        // S3上の画像のURLを取得
+                        $s3Url = Storage::disk('s3')->url($filePath);
 
-                    $image->file_path = $trimedFilePath;
+                        // 以前のpathを保管しておく
+                        $previousFilePath = $image->file_path;
+                        
+                        // 更新用の新しいurlを格納
+                        $image->file_path = $s3Url;
+
+                    } else {
+                        // 失敗時
+                        // 一時保存していたファイルを削除
+                        unlink($fileLocationFullPath);
+
+                        throw new Exception('s3への画像アップロードに失敗しました');
+                    }
                 }
             }
 
             $image->title = $validated_request['title'];
             $image->maker_id = $validated_request['maker_id'];
 
-            if($image->save()) {
+            if ($image->save()) {
+                if($request->file('file') && $s3Url) {
+                    // s3上の以前の画像を削除(strstr()はurlの整形)
+                    $trimedFilePath = strstr($previousFilePath, 'images');
+                    Storage::disk('s3')->delete($trimedFilePath);
+                    unlink($fileLocationFullPath);
+                }
+                
                 $maker = Maker::find($image->maker_id);
 
                 return response()->json([
                     'id' => $image['id'],
-                    'file_path' => Storage::url($image['file_path']),
+                    'file_path' => $image['file_path'],
                     'title' => $image['title'],
                     'maker' => $maker
                 ], 200);
+            } else {
+                throw new Exception('画像情報更新に失敗しました');
             }
         } catch (\Throwable $e) {
             \Log::error($e);
+
+            // s3に更新アップロードしていた場合は削除してclean up
+            if($request->file('file') && $s3Url && Storage::disk('s3')->exists($trimedFilePath)) {
+                Storage::disk('s3')->delete($trimedFilePath);
+
+                unlink($fileLocationFullPath);
+            }
 
             throw $e;
         }
@@ -201,11 +202,18 @@ class RacketImageController extends Controller
         try {
             $image = RacketImage::findOrFail($id);
 
-            Storage::disk('public')->delete($image->file_path);
+            // file_pathが下記の様に登録してある
+            // https://strii-bucket.s3.ap-northeast-1.amazonaws.com/images/rackets/20240313003254〜〜〜.jpg
+            // これを『images/rackets/20240313003254〜〜〜.jpg』の様に整形する
+            $trimedFilePath = strstr($image->file_path, 'images');
 
-            $image->delete();
+            if (Storage::disk('s3')->delete($trimedFilePath)) {
+                $image->delete();
 
-            return response()->json("{$image['title']}の画像を削除しました", 200);
+                return response()->json("{$image['title']}の画像を削除しました", 200);
+            } else {
+                throw new Exception("画像の削除に失敗しました");
+            }
         } catch (ModelNotFoundException $e) {
             throw $e;
         } catch (\Throwable $e) {
